% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_possible}
\alias{check_possible}
\title{Check input is possible}
\usage{
check_possible(type, possible, input)
}
\arguments{
\item{type}{A character string}

\item{possible}{A vector of characters or numerics}

\item{input}{A vector of characters or numerics}
}
\description{
Test whether the 'input' is among the possible/allowed entries.
}
