% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_content}
\alias{format_content}
\title{Format the content of a response.}
\usage{
format_content(resp, format = "tsv")
}
\arguments{
\item{resp}{An object of class \code{response}.}

\item{format}{A \code{character}. Only 'tsv' is used.}
}
\value{
A \code{tibble}.
}
\description{
Capture the API response in text format then reformat into a tibble.
}
\examples{
\dontrun{
# send request
url <- 'https://string-db.org/api/tsv/network?identifiers=PTCH1'
send_request(url)

# format content
format_content(resp, format = 'tsv')
}

}
