% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_query}
\alias{make_query}
\title{Make an API query}
\usage{
make_query(request, parameters, query_map)
}
\arguments{
\item{request}{A \code{character} string. Name of a valid request for an API}

\item{parameters}{A named \code{list} of the parameters for the query.}

\item{query_map}{A \code{data.frame} of 7 columns. For an example
\link{query_map_string}}
}
\value{
A \code{character} string of the formated query.
}
\description{
Check and format the query parameters according to the query_map.
}
\examples{
# query string API for a get_string_ids request
make_query(request = 'get_string_ids',
           parameters = list(identifiers = 'p53'),
           query_map = query_map_string)

}
