% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_maps.R
\docType{data}
\name{query_map_bioware}
\alias{query_map_bioware}
\title{query_map_bioware}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 7 columns.}
\usage{
query_map_bioware
}
\description{
query_map_bioware
}
\keyword{datasets}
