% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_request}
\alias{send_request}
\title{Send GET request}
\usage{
send_request(url)
}
\arguments{
\item{url}{A \code{character} string.}
}
\value{
An object of class \code{response}.
}
\description{
Send a GET request and return the response or the error.
}
\examples{
\dontrun{
# send request
url <- 'https://string-db.org/api/tsv/network?identifiers=PTCH1'
send_request(url)
}

}
